+++ CHANGELOG.md
@@ -43,6 +43,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
     - TV Shows
     - TV Seasons
     - Artists
+    - Movies
+    - Episodes
+    - Music Videos
+    - Other Videos
+    - Songs
+    - Images
   - Playlists can be added to schedules as a schedule item
   - Each time through the playlist, one item will be scheduled from each playlist item
     - NB: This does not mean every collection will always schedule one item; the normal flood playout restrictions like duration and fixed start times still apply here